pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "bili2text"
version = "0.3.0a1"
description = "CLI-first Bilibili to text toolkit with optional web and server features."
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
authors = [{ name = "Lanbin" }]
dependencies = [
  "typer>=0.12.5",
  "yt-dlp>=2025.3.31",
]

[project.optional-dependencies]
whisper = [
  "openai-whisper>=20240930",
]
sensevoice = [
  "funasr-onnx>=0.4.0",
  "jieba>=0.42.1",
  "torch>=2.5.0",
]
volcengine = [
  "requests>=2.32.3",
]
web = [
  "fastapi>=0.115.12",
  "jinja2>=3.1.6",
  "python-multipart>=0.0.20",
  "uvicorn>=0.34.0",
]
server = [
  "fastapi>=0.115.12",
  "jinja2>=3.1.6",
  "python-multipart>=0.0.20",
  "uvicorn>=0.34.0",
]
window = []

[project.scripts]
bili2text = "b2t.cli:main"

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/b2t"]
